+++ src/Krystal/Security/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.3
 ---
 
+ * Added `Filter::charsDecode()` wrapper to decode characters back
  * Added Crypter tool
  * Added `Filter::sanitize()` which can filter values by constants from `Sanitizeable` interface